Fix PathItem::resolveReferences() for array fields
The previous code `$this->$attribute[$k] = $referencedObject` had
no effect  if the value of `$attribute` is an array (as for instance
in the case of `parameters`), because `$this->$attribute` invokes
`SpecBaseObject::__get()`, which does *not* return a reference. Indeed
PHP issued `Notice: Indirect modification of overloaded property
cebe\openapi\spec\PathItem::$parameters has no effect` in this case.